Refactor interpreter module registry into focused submodules
